feature-matcher/requirements.txt
Flask==2.2.5
numpy==1.23.5
opencv_python==4.9.0.80
Pillow==10.4.0
Requests==2.32.3
torch==2.2.2
